Updated job filtering just the UI (#27)
I understand! The JavaScript code I added is perfect as it stands - it
only works with the data that's already rendered in the HTML template on
the page. It doesn't make any API calls or fetch data from external
sources. from claude
so hopefully this is just front end as ollie has created a API