Frontend connected to the API
This has taken me 2 hours and Claude has made some changes that broke it so i had to go on and try and amend it. There may be lots of extra code that isn't needed so I do apologise. It says that it is pulling from the backend however, I do not have the database yet so it will eventually get data from there.